fin_bot_v_0_1_3

+++ transaction_repository.py
@@ -3,13 +3,18 @@ from account import Account
 from database import get_connection
 from money import to_minor_units, from_minor_units
 from datetime import date
-from account_repository import AccountRepository
 
 class TransactionRepository:
     """Описание работы с таблицей transactions"""
 
-    def add_transaction(self, transaction : Transaction) -> int:
+    def add_transaction(self, transaction : Transaction, user_id : int) -> int:
         """Функция добавляет в таблицу transactions новую транзакцию"""
+
+        if type(user_id) is not int:
+            raise TypeError("Передаваемый идентификатор пользователя должен быть целочисленным")
+                                
+        if user_id <= 0:
+            raise ValueError("Передаваемый идентификатор пользователя должен быть больше нуля")
 
         if not isinstance(transaction, Transaction):
             raise TypeError("Должен быть передан объект Transaction")
@@ -21,6 +26,9 @@ class TransactionRepository:
             raise ValueError("Счёт транзакции сначала должен быть сохранен в базе")
 
         amount_minor = to_minor_units(transaction.amount)
+
+        if transaction.account.user_id != user_id:
+            raise ValueError("Счёт транзакции не принадлежит пользователю")
 
         connection = get_connection()
 
@@ -38,17 +46,36 @@ class TransactionRepository:
                     transfer_id,
                     is_active
                 )
-                VALUES (?,?,?,?,?,?,?,?)
+                SELECT 
+                    ?,
+                    ?,
+                    ?,
+                    ?,
+                    a.id,
+                    ?,
+                    ?,
+                    ?
+                FROM accounts AS a
+
+                WHERE a.id = ?
+                AND a.user_id = ?
+                
                 """,
-                (transaction.action_date.isoformat(),
-                amount_minor,
-                transaction.operation.value,
-                transaction.category,
-                transaction.account.object_number,
-                transaction.comment,
-                transaction.transfer_id,
-                int(transaction.is_active))
+                (
+                    transaction.action_date.isoformat(),
+                    amount_minor,
+                    transaction.operation.value,
+                    transaction.category,
+                    transaction.comment,
+                    transaction.transfer_id,
+                    int(transaction.is_active),
+                    transaction.account.object_number,
+                    user_id,
+                )
             )
+
+            if cursor.rowcount == 0:
+                raise ValueError("Счёт не найден или не принадлежит пользователю")
 
             connection.commit()
 
@@ -63,14 +90,20 @@ class TransactionRepository:
         finally:
             connection.close()
 
-    def delete_transaction_by_id(self,transaction_id :int) -> None:
+    def delete_transaction_by_id(self,transaction_id :int, user_id : int) -> None:
         """Функция удаляет транзакцию из базы данных"""
+
+        if type(user_id) is not int:
+            raise TypeError("Передаваемый идентификатор пользователя должен быть целочисленным")
+                        
+        if user_id <= 0:
+            raise ValueError("Передаваемый идентификатор пользователя должен быть больше нуля")
 
         if type(transaction_id) is not int:
             raise TypeError("Идентификатор транзакции должен быть целочисленного типа")
 
         if transaction_id <= 0:
-                raise ValueError("Идентификатор транзакции должен быть больше нуля")
+            raise ValueError("Идентификатор транзакции должен быть больше нуля")
 
         connection = get_connection()
 
@@ -79,8 +112,17 @@ class TransactionRepository:
                 """
                 DELETE FROM transactions
                 WHERE id = ?
+                AND transfer_id IS NULL
+                AND account_id IN (
+                    SELECT id
+                    FROM accounts
+                    WHERE user_id = ?
+                )
                 """,
-                (transaction_id,)
+                (
+                    transaction_id,
+                    user_id,
+                )
             )
 
             if cursor.rowcount == 0:
@@ -96,8 +138,14 @@ class TransactionRepository:
         finally:
             connection.close()
 
-    def update_transaction(self, transaction: Transaction) -> None:
+    def update_transaction(self, transaction: Transaction, user_id : int) -> None:
         """Функция обновляет данные существующей транзакции в таблице"""
+
+        if type(user_id) is not int:
+            raise TypeError("Передаваемый идентификатор пользователя должен быть целочисленным")
+                
+        if user_id <= 0:
+            raise ValueError("Передаваемый идентификатор пользователя должен быть больше нуля")
 
         if not isinstance(transaction, Transaction):
             raise TypeError("Должен быть передан объект Transaction")
@@ -107,6 +155,9 @@ class TransactionRepository:
 
         if transaction.account.object_number is None:
             raise ValueError("Счёт транзакции должен быть сохранён в базе")
+
+        if transaction.account.user_id != user_id:
+            raise ValueError("Счёт транзакции не принадлежит пользователю")
 
         amount_minor = to_minor_units(transaction.amount)
 
@@ -127,6 +178,21 @@ class TransactionRepository:
                     transfer_id = ?,
                     is_active = ?
                 WHERE id = ?
+                AND transfer_id IS NULL
+
+                AND EXISTS (
+                    SELECT 1 
+                    FROM accounts AS current_account
+                    WHERE current_account.id = transactions.account_id
+                    AND current_account.user_id = ?
+                )
+
+                AND EXISTS (
+                    SELECT 1
+                    FROM accounts AS new_account
+                    WHERE new_account.id = ?
+                    AND new_account.user_id = ?
+                )
                 """,
                 (
                     transaction.action_date.isoformat(),
@@ -137,7 +203,10 @@ class TransactionRepository:
                     transaction.comment,
                     transaction.transfer_id,
                     int(transaction.is_active),
-                    transaction.transaction_id
+                    transaction.transaction_id,
+                    user_id,
+                    transaction.account.object_number,
+                    user_id,
                 )
             )
 
@@ -152,67 +221,104 @@ class TransactionRepository:
         finally:
             connection.close()
 
-    def get_transaction_by_id(self, transaction_id : int) -> Transaction:
+    def get_transaction_by_id(self, transaction_id : int, user_id : int) -> Transaction | None:
         """Функция восстанавливает объект Transaction по идентификатору"""
 
         if type(transaction_id) is not int:
-            raise TypeError("Передаваемый идентификатор должен быть целочисленным")
+            raise TypeError("Передаваемый идентификатор транзакции должен быть целочисленным")
 
         if transaction_id <= 0:
-            raise ValueError("Передаваемый идентификатор должен быть больше нуля")
+            raise ValueError("Передаваемый идентификатор транзакции должен быть больше нуля")
+
+        if type(user_id) is not int:
+            raise TypeError("Передаваемый идентификатор пользователя должен быть целочисленным")
+        
+        if user_id <= 0:
+            raise ValueError("Передаваемый идентификатор пользователя должен быть больше нуля")
 
         connection = get_connection()
 
         try:
 
-            transaction_row = connection.execute(
+            row = connection.execute(
                 """
                 SELECT
-                    id,
-                    action_date,
-                    amount_minor,
-                    operation,
-                    category,
-                    account_id,
-                    comment,
-                    transfer_id,
-                    is_active
-                FROM transactions
-                WHERE id = ?
+                    t.id AS transaction_id,
+                    t.action_date,
+                    t.amount_minor,
+                    t.operation,
+                    t.category,
+                    t.comment,
+                    t.is_active AS transaction_is_active,
+                    t.transfer_id,
+
+                    a.id AS account_id,
+                    a.user_id,
+                    a.source,
+                    a.acc_type,
+                    a.product_name,
+                    a.requisites,
+                    a.currency,
+                    a.balance_minor,
+                    a.limit_minor,
+                    a.is_active AS account_is_active
+                
+                FROM transactions AS t
+
+                JOIN accounts AS a
+                    ON a.id = t.account_id
+
+                WHERE t.id = ?
+                AND a.user_id = ?
                 """,
-                (transaction_id,)
+                (
+                    transaction_id,
+                    user_id,
+                )
             ).fetchone()
 
-            if transaction_row is None:
-                raise ValueError("Не существует транзакции с таким идентификатором")
+            if row is None:
+                return None
 
         finally:
 
             connection.close()
 
-        account_repository = AccountRepository()
+        account = Account(
+            object_number= row["account_id"],
+            user_id= row["user_id"],
+            source= row["source"],
+            acc_type= row["acc_type"],
+            product_name= row["product_name"],
+            requisites= row["requisites"],
+            balance= from_minor_units(row["balance_minor"]),
+            currency= row["currency"],
+            limit= (from_minor_units(row["limit_minor"]) if row["limit_minor"] is not None else None),
+            is_active= bool(row["account_is_active"])
+        )
 
-        account = account_repository.get_account_by_id(transaction_row["account_id"])
-
-        if account is None:
-            raise ValueError("Счёта с указанным идентификатором не существует")
-
-        amount = from_minor_units(transaction_row["amount_minor"])
+        amount = from_minor_units(row["amount_minor"])
 
         return Transaction(
-                action_date= date.fromisoformat(transaction_row["action_date"]),
+                action_date= date.fromisoformat(row["action_date"]),
                 amount= amount,
-                operation= OperationType(transaction_row["operation"]),
-                category= transaction_row["category"],
+                operation= OperationType(row["operation"]),
+                category= row["category"],
                 account= account,
-                comment= transaction_row["comment"],
-                transaction_id= transaction_row["id"],
-                transfer_id= transaction_row["transfer_id"],
-                is_active= bool(transaction_row["is_active"])
+                comment= row["comment"],
+                transaction_id= row["transaction_id"],
+                transfer_id= row["transfer_id"],
+                is_active= bool(row["transaction_is_active"])
             )
 
-    def get_transactions_by_period(self, date_start: date, date_end: date, limit : int = 50, offset : int = 0) -> list[Transaction]:
+    def get_transactions_by_period(self, user_id : int, date_start: date, date_end: date, limit : int = 50, offset : int = 0) -> list[Transaction]:
         """Функция возвращает список транзакций за временной период"""
+
+        if type(user_id) is not int:
+            raise TypeError("Передаваемый идентификатор пользователя должен быть целочисленным")
+                
+        if user_id <= 0:
+            raise ValueError("Передаваемый идентификатор пользователя должен быть больше нуля")
 
         if not isinstance(date_start, date):
             raise TypeError("начальная дата должна быть объектом date")
@@ -252,6 +358,7 @@ class TransactionRepository:
                     t.is_active AS transaction_is_active,
 
                     a.id AS account_id,
+                    a.user_id AS user_id,
                     a.source AS account_source,
                     a.acc_type AS account_type,
                     a.product_name AS account_product_name,
@@ -268,6 +375,7 @@ class TransactionRepository:
 
                 WHERE t.action_date >= ?
                 AND t.action_date <= ?
+                AND a.user_id = ?
 
                 ORDER BY
                     t.action_date DESC,
@@ -279,6 +387,7 @@ class TransactionRepository:
                 (
                     date_start.isoformat(),
                     date_end.isoformat(),
+                    user_id,
                     limit,
                     offset,
                  )
@@ -294,6 +403,7 @@ class TransactionRepository:
 
             account = Account(
                 object_number=row["account_id"],
+                user_id= row["user_id"],
                 source= row["account_source"],
                 acc_type= row["account_type"],
                 product_name= row["account_product_name"],
